pyproject.toml
[tool.poetry]
name = "ph8"
version = "0.2.0"
description = ""
authors = ["Alex Howard <thezanke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
openai = "^0.28.0"
python-dotenv = "^1.0.0"
certifi = "^2023.7.22"
watchdog = {version = "^3.0.0", extras = ["watchmedo"]}
beautifulsoup4 = "^4.12.2"
tenacity = "^8.2.3"
langchain = "^0.0.308"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "ph8.main:init"
watch = "scripts.start:watch"
